+++ scripts/install_nvim.sh
@@ -2,16 +2,23 @@
 
 app_name="nvim"
 
+OS="$(uname)"
+if [[ "$OS" == "Darwin" ]]; then
+  RCFILE="$HOME/.zshrc"
+else
+  RCFILE="$HOME/.bashrc"
+fi
+
 # Make sure that .local/bin/ is in PATH
 line_to_append="export PATH=\"~/.local/bin:\$PATH\""
-# Check if the line already exists in .bashrc
-if grep -qF "$line_to_append" ~/.bashrc; then
+# Check if the line already exists in the shell rc file
+if grep -qF "$line_to_append" "$RCFILE"; then
   echo "~/.local/bin already in PATH"
 else
-  # If the line doesn't exist, append it to the end of .bashrc
-  echo "$line_to_append" >> ~/.bashrc
-  echo "~/.local/bin has been appended to ~/.bashrc."
-  source ~/.bashrc
+  # If the line doesn't exist, append it to the end of the shell rc file
+  echo "$line_to_append" >> "$RCFILE"
+  echo "~/.local/bin has been appended to $RCFILE."
+  source "$RCFILE"
 fi
 
 # Check if Neovim is installed
@@ -20,13 +27,17 @@ if which "$app_name" >/dev/null; then
 else
   # If the application is not found, install it
   echo "Neovim is not installed. Installing..."
-  neovim_url="https://github.com/neovim/neovim/releases/download/stable/nvim-linux-x86_64.appimage"
-  install_dir="$HOME/.local/bin"
-  mkdir -p "$install_dir"
-  neovim_appimage="$install_dir/nvim.appimage"
-  wget -qO "$neovim_appimage" "$neovim_url"
-  chmod +x "$neovim_appimage"
-  mv "$neovim_appimage" "$install_dir/nvim"
+  if [[ "$OS" == "Darwin" ]]; then
+    brew install neovim
+  else
+    neovim_url="https://github.com/neovim/neovim/releases/download/stable/nvim-linux-x86_64.appimage"
+    install_dir="$HOME/.local/bin"
+    mkdir -p "$install_dir"
+    neovim_appimage="$install_dir/nvim.appimage"
+    wget -qO "$neovim_appimage" "$neovim_url"
+    chmod +x "$neovim_appimage"
+    mv "$neovim_appimage" "$install_dir/nvim"
+  fi
   echo "Installation complete."
 fi
 
@@ -36,15 +47,19 @@ if which "rg" >/dev/null; then
 else
   # If the application is not found, install it
   echo "Ripgrep is not installed. Installing..."
-  ripgrep_url="https://github.com/BurntSushi/ripgrep/releases/download/14.1.0/ripgrep-14.1.0-x86_64-unknown-linux-musl.tar.gz"
-  install_dir="$HOME/.local/bin"
-  mkdir -p "$install_dir"
-  rg_tar="ripgrep.tar.gz"
-  wget -qO "$rg_tar" "$ripgrep_url"
-  tar -xvf "ripgrep.tar.gz"
-  mv "ripgrep-14.1.0-x86_64-unknown-linux-musl/rg" "$install_dir"
-  rm "ripgrep.tar.gz"
-  rm -rf "ripgrep-14.1.0-x86_64-unknown-linux-musl/rg"
+  if [[ "$OS" == "Darwin" ]]; then
+    brew install ripgrep
+  else
+    ripgrep_url="https://github.com/BurntSushi/ripgrep/releases/download/14.1.0/ripgrep-14.1.0-x86_64-unknown-linux-musl.tar.gz"
+    install_dir="$HOME/.local/bin"
+    mkdir -p "$install_dir"
+    rg_tar="ripgrep.tar.gz"
+    wget -qO "$rg_tar" "$ripgrep_url"
+    tar -xvf "ripgrep.tar.gz"
+    mv "ripgrep-14.1.0-x86_64-unknown-linux-musl/rg" "$install_dir"
+    rm "ripgrep.tar.gz"
+    rm -rf "ripgrep-14.1.0-x86_64-unknown-linux-musl/rg"
+  fi
   echo "Installation complete."
 fi
 
@@ -54,8 +69,8 @@ if which "node" >/dev/null; then
 else
   echo "Node.js is not installed. Installing node and npm..."
   # If the application is not found, install it
-  wget -qO- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.7/install.sh | bash
-  source ~/.bashrc
+  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.7/install.sh | bash
+  source "$RCFILE"
   nvm install node
   echo "Installation complete."
 fi
